Render Refactor 1
